config: abstract out create section from key logic
Abstracts out the logic for creating string buffer from given key for
example - 'branch.b' and returns '[branch "b"]'.

We want to keep the original config section intact in case of copy
operation. For this we need to fetch the section with updated new branch
name so that we can write that to the config file.

Signed-off-by: Sahil Dua <sahildua2305@gmail.com>